vezoon/requirements.txt
Django==2.2.11
django-configurations==2.2
django-environ==0.4.5
python-dotenv==0.10.5
django_debug_toolbar==2.2
djangorestframework==3.11.0
factory-boy==2.12.0
pytest-django==3.8.0
django-rest-swagger==2.2.0

